feat: fixed main class and methods
